test(core): split browserslist runtime polyfill tests
